requirements.txt
flask[async]
python-rapidjson
aiohttp